+++ packages/acp-bridge/package.json
@@ -51,6 +51,10 @@
       "types": "./dist/bridgeOptions.d.ts",
       "import": "./dist/bridgeOptions.js"
     },
+    "./spawnChannel": {
+      "types": "./dist/spawnChannel.d.ts",
+      "import": "./dist/spawnChannel.js"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
