+++ packages/acp-bridge/src/index.ts
@@ -13,3 +13,4 @@ export * from './status.js';
 export * from './bridgeErrors.js';
 export * from './bridgeTypes.js';
 export * from './bridgeOptions.js';
+export * from './spawnChannel.js';
